refactor(detections): extract ReDoS probe battery into a reusable module
Adds checkRuleTiming, the runtime pre-flight primitive the next tasks
wire into rule registration for pulled/custom pack rules. The CI gate
for bundled rules (redos.test.ts) is unchanged in behavior.
